Update index.html , style.css and app.js
I updated the "printUserAgeFunc" for printing the details about user's age and life days and after that i had to create some new divs or peragraphs tags for printing user's details. for doing this whole work i had to make changes in index.html and style.css so i did that.